Issue :
GitHub Pages - Failed to load resource: the server responded with a status of 404 ()

Solution :

copy index.html to 404.html

Reference :

https://router.vuejs.org/zh/guide/essentials/history-mode.html#%E5%90%8E%E7%AB%AF%E9%85%8D%E7%BD%AE%E4%BE%8B%E5%AD%90

https://router.vuejs.org/guide/essentials/history-mode.html#example-server-configurations

404 when reloading a Vue website published to Github pages

https://stackoverflow.com/questions/48521177/404-when-reloading-a-vue-website-published-to-github-pages/48522254

As a workaround, i duplicated the index.html and renamed it to 404.html.

In this way, if the page is reloaded, you still get the correct page however this is served through the 404.html file.

佈署 vue 到 GitHub Pages：404 頁面

https://siddharam.com.tw/post/20190929/

VUE CLI 打包build项目, 发布到github pages ,子路由刷新报错404 [ vue-router 里 mode: 'history' ]解决方法

https://blog.csdn.net/gentlu/article/details/99645232